deploy/requirements.txt
langgraph
typing-extensions
ormsgpack
langchain-core>=0.1.0
google-cloud-storage>=2.10.0
google-cloud-pubsub>=2.18.0
google-cloud-aiplatform>=1.36.0
google-cloud-firestore>=2.13.0
# google-cloud-sql-connector>=1.2.0
sqlalchemy>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
redis>=5.0.0
google-cloud-bigquery
optuna
matplotlib
seaborn 
